+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository._
 
+# Unreleased
+
+- Add modal to retrieve Slurm cluster queue using the `/job/squeue/` endpoint (\#1227).
+
 # 1.29.4
 
 - Expose core-related filtering/ordering options in task list (\#1225).